resistor-colors: Rename to resistor-color-duo
See https://github.com/exercism/problem-specifications/issues/1495
