+++ config.json
@@ -221,7 +221,7 @@
         "math"
       ]
     },
-    { "slug": "resistor-colors",
+    { "slug": "resistor-color-duo",
       "uuid": "57b0c57e-26a5-4ccf-aaf2-2fefddd918c1",
       "core": false,
       "unlocked_by": null,
