+++ exercises/resistor-color-duo/.meta/generator/resistor_color_duo_case.rb
@@ -1,0 +1,7 @@
+require 'generator/exercise_case'
+
+class ResistorColorDuoCase < Generator::ExerciseCase
+  def workload
+    assert_equal(expected, "ResistorColorDuo.value(#{colors})")
+  end
+end
