+++ exercises/resistor-color-duo/.meta/solutions/resistor_color_duo.rb
@@ -1,4 +1,4 @@
-module ResistorColors
+module ResistorColorDuo
   COLORS = %w(
     black
     brown
