--- exercises/resistor-colors/.meta/generator/resistor_colors_case.rb
@@ -1,7 +1,0 @@
-require 'generator/exercise_case'
-
-class ResistorColorsCase < Generator::ExerciseCase
-  def workload
-    assert_equal(expected, "ResistorColors.value(#{colors})")
-  end
-end
